Added my solution for 02-margin-and-padding-2
Some of the values are slightly different from the actual solution
that is provided already, but my solution is very similar.
I notice that they are more specific about which sides have margins and
padding whereas I am not as specific (For example, I use margin
whereas the solution uses margin-bottom).